82: Refactor batch create invoice new
