+++ app/javascript/api/invoices.js
@@ -2,7 +2,7 @@ import axios from './axios.js'
 
 export default {
   loadInvoiceForm (params, cb) {
-    axios.get('/invoices/new.json', {
+    axios.get('/batch_create/invoices/new.json', {
       params: params
     }).then(function(response) {
       cb(response.data)
@@ -11,7 +11,7 @@ export default {
     });
   },
   createInvoices (params, cb) {
-    axios.post('/invoices.json',
+    axios.post('/batch_create/invoices.json',
       { invoices: { multi_invoices: params } },
       { headers: {'X-CSRF-TOKEN': document.querySelector('meta[name="csrf-token"]').getAttribute('content')}}
     ).then(function(response) {
