+++ app/javascript/packs/application.js
@@ -25,6 +25,7 @@ import NewUser from '../newUser.vue'
 import Floors from '../floor.vue'
 import Contracts from '../contract.vue'
 import Invoices from '../invoice.vue'
+import BatchCreateNewInvoices from '../batchCreateNewInvoice.vue'
 
 Vue.use(FlashMessage)
 Vue.use(VueI18n)
@@ -39,6 +40,7 @@ document.addEventListener('DOMContentLoaded', () => {
   const elementFloor = document.getElementById('floorApp')
   const elementContract = document.getElementById('contractApp')
   const elementInvoice = document.getElementById('invoiceApp')
+  const elementBatchCreateNewInvoiceApp = document.getElementById('batchCreateNewInvoiceApp')
   const i18n = new VueI18n({
     locale: locale,
     messages: loadLocaleMessages(),
@@ -80,12 +82,12 @@ document.addEventListener('DOMContentLoaded', () => {
     })
   }
 
-  if(elementInvoice != null) {
-    var invoiceApp = new Vue({
-      el: '#invoiceApp',
+  if(elementBatchCreateNewInvoiceApp != null) {
+    var batchCreateNewInvoiceApp = new Vue({
+      el: '#batchCreateNewInvoiceApp',
       store,
       i18n,
-      render: h => h(Invoices, {})
+      render: h => h(BatchCreateNewInvoices, {})
     })
   }
 })
