+++ app/views/batch_create/invoices/new.html.slim
@@ -1,3 +1,3 @@
 .card.h-full.mb-2
   .card-body
-    #newInvoiceApp
+    #batchCreateNewInvoiceApp
